refactor: extract vsock host command core
